fix: add exchangeRateDailyRepository && separate responsibility
